back/app/requirements.txt
aiohttp==3.8.5
aiopg==1.4.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.103.1
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2